backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
PyYAML==6.0.3
python-chess==1.999
httpx==0.28.1
